scripts/requirements.txt
basketball_reference_web_scraper==4.15.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
greenlet==3.1.1
idna==3.10
lxml==5.3.0
nba-headshot-downloader==1.0.0
nba_api==1.7.0
numpy==1.26.4
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
